+++ config/environment.rb
@@ -6,7 +6,6 @@ require 'i_am'
 require 'date_time_utilities'
 require 'active_job/after_commit_runner'
 require 'active_record/indestructible_record'
-require 'active_record/background_migration'
 require 'acts_as_resource'
 require 'acts_as_tasked'
 require 'belongs_to_time_zone'
