+++ db/migrate/20180223002900_add_question_index_to_tasks_tasked_exercises.rb
@@ -1,0 +1,10 @@
+class AddQuestionIndexToTasksTaskedExercises < ActiveRecord::Migration
+  def change
+    add_column :tasks_tasked_exercises, :question_index, :integer
+
+    reversible do |dir|
+      dir.up   { BackgroundMigrate.perform_later 'up',   20180219163045 }
+      dir.down { BackgroundMigrate.call          'down', 20180219163045 }
+    end
+  end
+end
